mplus/no_2/app/requirements.txt
asgiref==3.3.4
click==8.0.1
dnspython==2.1.0
fastapi==0.65.2
h11==0.12.0
pydantic==1.8.2
pymongo==3.11.4
schematics==2.1.0
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
